// src/app/page/master/negara/negara.interface.ts
    import {EventEmitter} from '@angular/core';

    export interface Negara {
		kdProfile?: string;
		kdNegara?: string;
		kodeTelepon?: string;
		kdNegaraHead?: number;
		namaNegara?: string;
		reportDisplay?: string;
		kodeExternal?: string;
		namaExternal?: string;
		statusEnabled?: boolean;
		noRec?: string;
		kode?:any;
		version?:any;
    }